requirements.txt
PySide6>=6.5
paramiko>=3.0
pyte>=0.8
